+++ src/DataManager.ts
@@ -166,3 +166,10 @@ export async function getScheduleData(
     data: schedule(json),
   };
 }
+
+export async function clearCache() {
+  const cacheLocations = ['plan', 'schedule0', 'schedule1', 'schedule2'];
+  for (const loc of cacheLocations) {
+    await localforage.removeItem(loc);
+  }
+}